Switch to a full fork of segmentio/kafka-go (#5)
Golang's dependency management has never been very good, but as it
happens, the `replace` directive in a go.mod file is only honored
when building `main.go`. It's ignored when a dependency specifies
that a modulepath is to be replaced by a different module. This
would leak the in-between state of our dependency on kafka-go (one
where I'm submitting my changes upstream but it's taking time for
them to go in) to the user, which Would Be Bad. So instead this
hard fork prevents us from leaking details up the dependency chain.